feat: more none checking

+++ fishsense_lite/pipeline/tasks/detect_laser.py
@@ -14,23 +14,27 @@ def detect_laser(
     laser_detector: LaserDetector,
     debug_path: Path,
 ) -> np.ndarray:
+    if img is None or laser_detector is None:
+        return None
+
     laser_image_coords = laser_detector.find_laser(img)
 
     if laser_detection is None:
         return None
 
-    png_name = input_file.name.replace("ORF", "PNG").replace("orf", "png")
-    laser_detection_path = debug_path / f"detection_{png_name}"
-    if laser_detection_path.exists():
-        laser_detection_path.unlink()
+    if debug_path is not None:
+        png_name = input_file.name.replace("ORF", "PNG").replace("orf", "png")
+        laser_detection_path = debug_path / f"detection_{png_name}"
+        if laser_detection_path.exists():
+            laser_detection_path.unlink()
 
-    laser_detection = cv2.circle(
-        img_as_ubyte(img),
-        np.round(laser_image_coords).astype(int),
-        radius=5,
-        color=(0, 255, 0),
-        thickness=-1,
-    )
-    cv2.imwrite(laser_detection_path.absolute().as_posix(), laser_detection)
+        laser_detection = cv2.circle(
+            img_as_ubyte(img),
+            np.round(laser_image_coords).astype(int),
+            radius=5,
+            color=(0, 255, 0),
+            thickness=-1,
+        )
+        cv2.imwrite(laser_detection_path.absolute().as_posix(), laser_detection)
 
     return laser_image_coords
